ch4/ofi: Avoid warning when converting void * to uint64_t
The compiler will complain when converting a pointer to an integer of
different size. First cast to uintptr_t, which can always hold a
pointer value, then to uint64_t for use with libfabric.

+++ src/mpid/ch4/netmod/ofi/ofi_rma.h
@@ -830,7 +830,7 @@ static inline int MPIDI_NM_mpi_compare_and_swap(const void *origin_addr,
     resultv.count = 1;
     comparev.addr = (void *) compare_addr;
     comparev.count = 1;
-    targetv.addr = (uint64_t) tbuffer;
+    targetv.addr = (uint64_t) (uintptr_t) tbuffer;
     targetv.count = 1;
     targetv.key = MPIDI_OFI_winfo_mr_key(win, target_rank);;
 
@@ -1364,7 +1364,7 @@ static inline int MPIDI_NM_mpi_fetch_and_op(const void *origin_addr,
     originv.count = 1;
     resultv.addr = (void *) rbuffer;
     resultv.count = 1;
-    targetv.addr = (uint64_t) tbuffer;
+    targetv.addr = (uint64_t) (uintptr_t) tbuffer;
     targetv.count = 1;
     targetv.key = MPIDI_OFI_winfo_mr_key(win, target_rank);
